updated network error handling
The network problem should be resolved now, if the user is not connected to network, then they should connect and click "ok" to reload data. If in activity lost network again connect and click "ok" to reload the data and continue the activity in the app. It should not exit from the app and crash